Added BaseAddress as a base class for the Addresses funtionality
